requirements.txt
cs50
Flask
google-cloud-translate